fix: stabilize reader page loading
Extract the DIVINA page image pipeline into a dedicated scheduler so the reader state model no longer owns download, decode, upscale, and animated playback task coordination.

Prioritize visible pages, cancel stale work outside the active window, and keep the current page synchronized with the scheduler across view-state regeneration and scroll commits.

Also commit the centered item immediately when iOS resigns active and clear stale view items during book reload resets to avoid page jumps and leftover reader state after interrupted or failed loads.
